% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratifiedValidation.R
\name{addEvaluation}
\alias{addEvaluation}
\title{addEvaluation}
\usage{
addEvaluation(performanceEvaluation, subpopEvaluation)
}
\arguments{
\item{performanceEvaluation}{The main result performanceEvaluation}

\item{subpopEvaluation}{The subpopulation result}
}
\value{
An object of class \code{runPlp} that is recalibrated on the new data
}
\description{
Adds the subpopulation results to the main results
}
\details{
Append the subpopulation results into the main results
}
