% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovCoagFullValidation.R
\docType{package}
\name{CovCoagFullValidation}
\alias{CovCoagFullValidation}
\title{CovCoagFullValidation}
\description{
A package for validating prediction models using data in the OMOP CDM
}
