% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraPredictFunctions.R
\name{predict.nonPlpGlm}
\alias{predict.nonPlpGlm}
\title{Prediction for existing GLM}
\usage{
\method{predict}{nonPlpGlm}(plpModel, plpData, population)
}
\arguments{
\item{plpModel}{The model being applied}

\item{plpData}{The new data}

\item{population}{The new population}
}
\value{
The population with an extra column 'value' corresponding to the patients risk
}
\description{
Prediction for existing GLM
}
\details{
This applies the existing models and calcualtes the risk for a population
}
