% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraPredictFunctions.R
\name{predict.pythonJson}
\alias{predict.pythonJson}
\title{Prediction for python models saved as json objects}
\usage{
\method{predict}{pythonJson}(plpModel, plpData, population)
}
\arguments{
\item{plpModel}{The model being applied}

\item{plpData}{The new data}

\item{population}{The new population}
}
\value{
The population with an extra column 'value' corresponding to the patients risk
}
\description{
Prediction for python models saved as json objects
}
\details{
This applies the existing models and calculates the risk for a population
}
