% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratifiedValidation.R
\name{stratifiedValidation}
\alias{stratifiedValidation}
\title{stratifiedValidation}
\usage{
stratifiedValidation(prediction, analysisId)
}
\arguments{
\item{prediction}{the prediction object from the results object}

\item{analysisId}{The analysis identifier}
}
\value{
An object of class \code{runPlp} that is recalibrated on the new data
}
\description{
Evaluates PLP model in subpopulations
}
\details{
Append the subpopulation results into the main results
}
